remove old code related to LogEntry.ts
