--- src/inspect.ts
@@ -1,3 +1,0 @@
-import { logStorage } from "./logStorage";
-
-console.log(logStorage.getItem(Bun.argv[2]));
